[#32] refactor : Move the Coordinates interface to the location file
	기존의 kakaoMap.type.ts에 존재하던 Coordinates interface를
	location.type.ts파일로 옮겼습니다.
