+++ src/types/location.type.ts
@@ -1,0 +1,4 @@
+export interface Coordinates {
+  latitude: number;
+  longitude: number;
+}
